Add MEOrderBook unit tests + portability shims
Add a standalone unit-test suite for Exchange::MEOrderBook that drives
orders through a real MatchingEngine (single-threaded) and asserts on the
emitted client responses and market updates. Covers resting adds, full and
partial matches, non-crossing orders, FIFO price-time priority, best-price
selection, cancels, and cancel-rejections (39 checks).

Also add small, platform-guarded shims so the x86/Linux code compiles and
runs on arm64 macOS (the x86/Linux path is byte-for-byte unchanged):
- perf_utils.h: aarch64 (cntvct_el0) branch for rdtsc()
- thread_utils.h: no-op thread affinity off Linux
- lf_queue.h: cast pthread_self() for platforms where pthread_t is a pointer
- me_order_book.h: explicit -> Priority return type on getNextPriority()
  (fixes an ill-formed auto deduction where Priority != unsigned long)

tests/run_tests.sh builds under AddressSanitizer by default (the full-size
config reserves ~16GB of virtual memory which the plain macOS allocator
mishandles; ASAN's allocator handles it and adds memory-safety checks).
tests/README.md documents usage and the rationale.

+++ Chapter12/common/lf_queue.h
@@ -29,7 +29,7 @@ namespace Common {
 
     auto updateReadIndex() noexcept {
       next_read_index_ = (next_read_index_ + 1) % store_.size(); // wrap around at the end of container size.
-      ASSERT(num_elements_ != 0, "Read an invalid element in:" + std::to_string(pthread_self()));
+      ASSERT(num_elements_ != 0, "Read an invalid element in:" + std::to_string((uintptr_t) pthread_self()));
       num_elements_--;
     }
